agentcore-pydantic-ai/pyproject.toml
[project]
name = "agentcore-pydantic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=0.1.3",
    "bedrock-agentcore-starter-toolkit>=0.1.9",
    "pydantic-ai-slim[bedrock]>=1.0.1",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "memory"]

[dependency-groups]
dev = [
    "vulture>=2.14",
    "ruff>=0.8.0",
]
